# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Proof Fabric Protocol (PFP) API
  description: Cryptographically verifiable Financial Evidence Artifacts — enterprise
    control plane.
  version: 2.0.0
paths:
...
  /api/auth/login:
    post:
      tags:
      - Auth
      summary: Login
      operationId: login_api_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/me:
    get:
      tags:
      - Auth
      summary: Me
      operationId: me_api_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
...
  /api/auth/logout:
    post:
      tags:
      - Auth
      summary: Logout
      operationId: logout_api_auth_logout_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/admin/api-keys/create:
    post:
      tags:
      - Admin
      summary: Create Api Key
      operationId: create_api_key_api_admin_api_keys_create_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApiKeyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/admin/api-keys:
    get:
      tags:
      - Admin
      summary: List Api Keys
      operationId: list_api_keys_api_admin_api_keys_get
      security:
      - HTTPBearer: []
      parameters:
      - name: tenant_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/api-keys/revoke:
    post:
      tags:
      - Admin
      summary: Revoke Api Key
      operationId: revoke_api_key_api_admin_api_keys_revoke_post
      security:
      - HTTPBearer: []
      parameters:
      - name: key_id
        in: query
        required: true
        schema:
          type: string
          title: Key Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/keys/create:
    post:
      tags:
      - Admin
      summary: Create Signing Key
      description: Register an externally-generated public key in the registry.
      operationId: create_signing_key_api_admin_keys_create_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKeyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/admin/keys/rotate:
    post:
      tags:
      - Admin
      summary: Rotate Signing Key
      description: 'Generate a new keypair, register its public key as active, retire
        the

        previous active key(s). Returns the new private seed ONCE — the operator

        must deploy it to the KMS/secret store and restart for new issuance to use
        it.'
      operationId: rotate_signing_key_api_admin_keys_rotate_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/admin/keys/revoke:
    post:
      tags:
      - Admin
      summary: Revoke Signing Key
      operationId: revoke_signing_key_api_admin_keys_revoke_post
      security:
      - HTTPBearer: []
      parameters:
      - name: public_key_id
        in: query
        required: true
        schema:
          type: string
          title: Public Key Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/keys/retire:
    post:
      tags:
      - Admin
      summary: Retire Signing Key
      operationId: retire_signing_key_api_admin_keys_retire_post
      security:
      - HTTPBearer: []
      parameters:
      - name: public_key_id
        in: query
        required: true
        schema:
          type: string
          title: Public Key Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/tenants:
    get:
      tags:
      - Admin
      summary: List Tenants
      operationId: list_tenants_api_admin_tenants_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
    post:
      tags:
      - Admin
      summary: Create Tenant
      operationId: create_tenant_api_admin_tenants_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/admin/audit:
    get:
      tags:
      - Admin
      summary: Get Audit
      operationId: get_audit_api_admin_audit_get
      security:
      - HTTPBearer: []
      parameters:
      - name: tenant_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      - name: action
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Action
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 50
          title: Limit
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/audit/verify:
    get:
      tags:
      - Admin
      summary: Verify Audit
      description: Verify the integrity of the audit hash chain.
      operationId: verify_audit_api_admin_audit_verify_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    CreateApiKeyRequest:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
        customer_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Customer Id
        scopes:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Scopes
        expires_in_days:
          anyOf:
          - type: integer
            maximum: 3650.0
            minimum: 1.0
          - type: 'null'
          title: Expires In Days
      type: object
      required:
      - name
      title: CreateApiKeyRequest
    CreateApiKeyResponse:
      properties:
        key_id:
          type: string
          title: Key Id
        api_key:
          type: string
          title: Api Key
        name:
          type: string
          title: Name
        tenant_id:
          type: string
          title: Tenant Id
        scopes:
          items:
            type: string
          type: array
          title: Scopes
        expires_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Expires At
      type: object
      required:
      - key_id
      - api_key
      - name
      - tenant_id
      - scopes
      title: CreateApiKeyResponse
    CreateKeyRequest:
      properties:
        public_key_b64:
          type: string
          minLength: 10
          title: Public Key B64
        public_key_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Public Key Id
      type: object
      required:
      - public_key_b64
      title: CreateKeyRequest
    CreateTenantRequest:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      type: object
      required:
      - name
      title: CreateTenantRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        role:
          type: string
          title: Role
        tenant_id:
          type: string
          title: Tenant Id
        email:
          type: string
          title: Email
      type: object
      required:
      - access_token
      - refresh_token
      - role
      - tenant_id
      - email
      title: TokenResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: https://pfp-evidence.preview.emergentagent.com
  description: Sandbox/preview
- url: https://api.pfprotocol.com
  description: Production